Correcciones menores.
- Cambio en el tipo de dato idUsuario
- Eliminacion de clases redundantes

+++ src/main/java/com/xtheggx/monedaDOS/service/TransaccionServiceImpl.java
@@ -175,7 +175,7 @@ public class TransaccionServiceImpl implements  TransaccionService{
                     break;
                 case "usuarioId":
                     if (valor instanceof Number num) {
-                        int nuevoUsuarioId = num.intValue();
+                        Long nuevoUsuarioId = num.longValue();
                         if (!tx.getUsuario().getIdUsuario().equals(nuevoUsuarioId)) {
                             var nuevoUsuario = usuarioRepo.findById(nuevoUsuarioId)
                                     .orElseThrow(() -> new ResourceNotFoundException(
